sesion_07/requirements.txt
Django~=3.2
graphene-django==2.15.0
